requirements.txt
aniso8601==9.0.1
annotated-types==0.7.0
asgiref==3.8.1
blinker==1.8.2
cachetools==4.2.4
certifi==2024.7.4
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
dnspython==2.6.1
email_validator==2.2.0
Flask==2.3.3
Flask-Bcrypt==1.0.1
Flask-JWT-Extended==4.6.0
Flask-WTF==1.2.1
google-auth==1.27.0
google-auth-oauthlib==0.4.2
httplib2==0.22.0
idna==3.7
itsdangerous==2.2.0
Jinja2==3.1.4
MarkupSafe==2.1.5
mongoengine==0.28.2
oauthlib==3.2.2
promise==2.3
pyasn1==0.6.0
pyasn1_modules==0.4.0
pydantic==2.8.2
pydantic_core==2.20.1
PyJWT==2.9.0
pymongo==4.8.0
pyparsing==3.1.4
python-dotenv==1.0.1
pytz==2024.1
requests==2.32.3
requests-oauthlib==2.0.0
rsa==4.9
Rx==1.6.3
six==1.16.0
typing_extensions==4.12.2
urllib3==2.2.2
URLObject==2.4.3
Werkzeug==3.0.4
WTForms==3.1.2

pip~=24.0
openai~=1.44.1